Move some files, remove unused files.
